interface/requirements.txt
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==9.5.0
fastparquet==2023.2.0
tensorflow==2.12.0
sklearn==0.0